Create YahtzeePro.Optimum.csproj and move Optimum Calculation classes to it
