+++ YahtzeePro.CalculateOptimum.Api/Program.cs
@@ -1,5 +1,7 @@
 using System.Text.Json;
 using YahtzeePro;
+using YahtzeePro.models;
+using YahtzeePro.Optimum;
 
 internal class Program
 {
