+++ YahtzeePro.CalculateOptimum.Api/YahtzeePro.CalculateOptimum.Api.csproj
@@ -7,6 +7,7 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\YahtzeePro.Optimum\YahtzeePro.Optimum.csproj" />
     <ProjectReference Include="..\YahtzeePro\YahtzeePro.csproj" />
   </ItemGroup>
