+++ YahtzeePro.Optimum/IOptimumCalculator.cs
@@ -1,4 +1,6 @@
-namespace YahtzeePro
+using YahtzeePro.models;
+
+namespace YahtzeePro.Optimum
 {
     public interface IOptimumCalculator
     {
