+++ YahtzeePro.Play/Game.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using YahtzeePro.models;
 
 namespace YahtzeePro.Play;
