+++ YahtzeePro.tests/OptimumCalculatorTests.cs
@@ -1,5 +1,7 @@
 using Microsoft.Extensions.Logging.Abstractions;
 using Xunit;
+using YahtzeePro.models;
+using YahtzeePro.Optimum;
 
 namespace YahtzeePro.tests
 {
@@ -10,7 +12,7 @@ namespace YahtzeePro.tests
         [Fact]
         public void SimpleCaseCheckSizeOfResults()
         {
-            var calculator = new OptimumCalculator(NullLogger<OptimumCalculator>.Instance);
+            var calculator = new OptimumCalculator(NullLogger<OptimumCalculator>.Instance); 
             var result = calculator.Calculate(_gameConfiguration);
             Assert.Equal(2, result.Count);
         }
