--- YahtzeePro/GameState.cs
@@ -1,12 +1,0 @@
-﻿namespace YahtzeePro
-{
-    [Serializable]
-    public readonly record struct GameState(
-        int PlayerScore, 
-        int OpponentScore, 
-        int CachedScore, 
-        int DiceToRoll, 
-        bool IsStartOfTurn, 
-        int TotalDice
-    );
-}
