+++ YahtzeePro/models/GameState.cs
@@ -1,0 +1,11 @@
+﻿namespace YahtzeePro.models;
+
+[Serializable]
+public readonly record struct GameState(
+    int PlayerScore, 
+    int OpponentScore, 
+    int CachedScore, 
+    int DiceToRoll, 
+    bool IsStartOfTurn, 
+    int TotalDice
+);
